Safari-611.1.10/LayoutTests/svg/animations/svgenum-animation-6-expected.txt

SVG 1.1 dynamic animation tests

Test ColorMatrixType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_SATURATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_SATURATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_HUEROTATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_HUEROTATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS successfullyParsed is true

TEST COMPLETE
